searchcore: keep the alignment score in 64 bits (struct hit::nwscore)
Every aligner computes the score of a hit in an int64_t, and every call
site then narrowed it into struct hit's int nwscore with a static_cast.
A score above 2^31 wrapped around: a 70 kb sequence aligned against
itself with --match 32767 printed raw -2001277296 instead of 2293690000,
in --search_exact and --usearch_global alike (the other two call sites,
cluster and allpairs_global, narrowed the same way).

nwscore is now int64_t and the four narrowing casts are gone. The size
of struct hit does not change: the field sits in front of a 4-byte hole
(before double nwid), so nwdiff, nwgaps, nwindels and nwalignmentlength
move by 4 bytes and everything from nwid onward keeps its offset
(sizeof 184 before and after, verified on x86_64 with offsetof and on
aarch64, powerpc64le, mips64el and x86_64-w64-mingw32 with
static_assert). Library users compiled against the old header would
read those four fields at the wrong offset: this is an ABI change.

The only reader is the raw userfield, which prints through the
fprint_integer template; nothing sorts, compares or filters on the
score, so the runtime cost is an 8-byte store instead of a 4-byte one,
once per aligned hit.

+++ src/commands/search_exact.cpp
@@ -215,7 +215,7 @@ auto add_hit(struct searchinfo_s * si, uint64_t const seqno) -> void
       hp->count = 0;
 
       auto const qseqlen = static_cast<int>(si->qsequence.size());
-      hp->nwscore = static_cast<int>(exact_match_score(View<char>{si->qsequence}, *si->parameters));
+      hp->nwscore = exact_match_score(View<char>{si->qsequence}, *si->parameters);
       hp->nwdiff = 0;
       hp->nwgaps = 0;
       hp->nwindels = 0;
